Extract command and plugin endpoints
