+++ src/TaskHub.Server/PluginEndpoints.cs
@@ -1,0 +1,13 @@
+using Microsoft.AspNetCore.Builder;
+using Microsoft.AspNetCore.Routing;
+
+namespace TaskHub.Server;
+
+public static class PluginEndpoints
+{
+    public static IEndpointRouteBuilder MapPluginEndpoints(this IEndpointRouteBuilder app)
+    {
+        app.MapGet("/dlls", (PluginManager plugins) => plugins.LoadedAssemblies);
+        return app;
+    }
+}
